Remaked solution strucrure
