+++ SkillAppAdoDapperWebApi.BLL/SkillAppAdoDapperWebApi.BLL.csproj
@@ -1,15 +1,12 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="EntityFramework" Version="6.4.4" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\SkillAppAdoDapperWebApi.DAL\SkillAppAdoDapperWebApi.DAL.csproj" />
+    <ProjectReference Include="..\SkillAppAdoDapperWebApi.Repository\SkillAppAdoDapperWebApi.Repository.csproj" />
   </ItemGroup>
 
 </Project>
