--- SkillAppAdoDapperWebApi.DAL/Interfaces/IConnectionFactory.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SkillManagement.DataAccess.Interfaces
-{
-    public interface IConnectionFactory
-    {
-        IDbConnection GetSqlConnection { get; }
-        void SetConnection(string connectionString);
-    }
-}
