+++ SkillAppAdoDapperWebApi.Repository/Interfaces/IGenericRepository.cs
