+++ SkillAppAdoDapperWebApi/Startup.cs
@@ -1,16 +1,9 @@
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 using Microsoft.EntityFrameworkCore;
-using SkillManagement.DataAccess.Infrastructure;
 using SkillManagement.DataAccess.Interfaces;
 using SkillManagement.DataAccess.Interfaces.SQLInterfaces.ISQLRepositories;
 using SkillManagement.DataAccess.Interfaces.SQLInterfaces.ISQLServices;
@@ -54,7 +47,6 @@ namespace SkillAppAdoDapperWebApi
 
             services.AddTransient<ISQLunitOfWork, SQLsqlunitOfWork>();
 
-            services.AddTransient<IConnectionFactory, ConnectionFactory>();
             services.AddSingleton<IConfiguration>(_Configuration);
 
             services.AddSwaggerGen(c =>
